requirements.txt
pymongo==4.8.0
python-dotenv==1.0.1
streamlit==1.39.0
pandas==2.2.3 